Return a sensible default GOVUK_APP_DOMAIN value
The way Plek works, and is used, means that it tries to find
`ENV['GOVUK_APP_DOMAIN']` immediately, which will causes any Rails apps
that include Plek in an initialiser to die when instantiating
themselves.

This makes simple tasks like `rake db:migrate` or `rails generate`
somewhat more painful to run.

This fixes the problem by only failing if the perceived RAILS/RACK_ENV
is production, otherwise it returns a sensible default.