Allow custom URIs for services to be defined
Allow custom URIs for services to be defined using environment
variables, so that it is easier to set up and run multiple
services in isolation without them having to be running at
particular hostnames.

A service URI can be defined using the following environment
variable naming convention:

    # replace <service> with the name of the service, uppercased
    # and underscored
    export PLEK_SERVICE_<service>_URI=http://localhost:3000

Existing behaviour of the gem is not affected when these
environment variables are not present.