Add a method that returns `URI` objects
Using `URI` objects instead of strings means we can build paths without
having to concatenate strings together and worrying about, for instance,
whether the URI ends with a slash or not. For example:

    Plek.find_uri("stuff") + "things"  # URL:http://stuff.dev.gov.uk/things

rather than:

    Plek.find("stuff") + "things"  # "http://stuff.dev.gov.ukthings"

The method name is for consistency with the `asset_uri` and
`website_uri` methods, so that any method ending `_uri` will return a
`URI` object, rather than a string.