Add a new `environment` property to Plek
Provides an easy way to identify the environment we’re currently
running under.

For backward compatibility purposes, what used to be “preview” (i.e.
`preview.alphagov.co.uk`) is now referred to as “integration” using
this API. This provides a stable value for any applications using this
new API without needing to be concerned about distinctions between
preview and the new integration environment.