Add public_asset_host method
This returns the GOVUK_ASSET_HOST, which is suitable for public
consumption. This will resolve to
https://assets.digital.cabinet-office.gov.uk in production.

This is mostly needed for Publishing apps like Whitehall to
differentiate between the Rails-style asset root used to serve JS and
CSS for Admin interfaces, versus the public CDN that they should
reference in content that is pushed to publishing-api.