Allow use of GOVUK_APP_DOMAIN_INTERNAL
In AWS, we use both internal and "public" DNS names in how we route
traffic. Traffic between services within the environment should always
be internal.

Applications use Plek for service discovery of internal services, but
also use it for self referred lookups for user facing redirects. This
causes an issue as we cannot set the default service discovery to use
the internal domains incase we send a user to the wrong URL.

In most places we set PLEK_SERVICE_APPLICATION_URI to ensure overrides
happen in the correct places.

However, content-store [uses a Plek lookup to help set a backend
route][1], and we cannot determine which specific service it will do
this for. I am hesitant to set the default GOVUK_APP_DOMAIN as the
internal domain as I do not know what side effects this may cause, but
we do know that Router backends should almost always refer to the
internal domain to route traffic.

We previously [set GOVUK_APP_DOMAIN_INTERNAL environment variable][2] to use
within Smokey, so we can make use of the this.

We can specifically ask Plek to use the internal domain, and only set it
if the appropriate environment variable exists.

[1]: https://github.com/alphagov/content-store/blob/3ea284672bc2ee15a6f17c9936847cbc7f605ea0/app/models/route_set.rb#L102
[2]: https://github.com/alphagov/govuk-puppet/commit/d7d231efde8b498c07c727ab448c20c507396749