Add support for explicit external routing
Through a new environment variable called GOVUK_APP_DOMAIN_EXTERNAL.

With the new routing in AWS, where the domains for internal and
external routes are different, Plek is not working so well for
routing. Configuring the external vs internal domain on a per service
basis will work, but the services using Plek know if they are looking
for an internal or external URL, so lets allow them to specify
this.

I see Plek as currently being primarilly used for internal routing,
but some external use cases have crept in, so this change should
handle these properly.