Use GOVUK_APP_DOMAIN when GOVUK_APP_DOMAIN_EXTERNAL isn't set
Introducing the need for this environment variable (in [1]) seems to have
caused a variety of pain points with situations where we don't have a
GOVUK_APP_DOMAIN_EXTERNAL set. Frequently we're seeing builds fail due
to the missing env var. In most cases where it doesn't exist
GOVUK_APP_DOMAIN is set to an external domain anyway.

This sets the env var to fallback to the parent domain if that is set
which should make this change now backwards compatible.

[1]: https://github.com/alphagov/plek/pull/55