Remove Plek.current method
This method was deprecated back in October 2012 [1]. This now removes
the method and resolves this todo.

It's unfortunate that there was not a warning for the use of this
method, but as it's been such a long time and that we're working towards
a new major version this feels appropriate.

[1]: https://github.com/alphagov/plek/commit/309ef5942a7adcdc6d68a9416fa3267c17a2268f