Remove URI based look up methods
These seemed unnecessary as it's just running URI over the returned
argument and clients can do that. These had fallen out of sync with the
interface the class actually had so it seemed best to remove them.

From my GitHub search there's only a couple of places these are used in
our code.