Reject malformed service names
Previously this tool was very forgiving in what user input it accepted
and this meant that some really horribly wrong stuff could be attempted
to be salvaged. For example:

```
irb(main):005:0> Plek.find(" S76I%Gn?O^n    ")
=> "http://signon.dev.gov.uk"
```

Which after this change now does:

```
irb(main):001:0> Plek.find(" S76I%Gn?O^n    ")
Traceback (most recent call last):
       16: from /Users/kevin.dew/.rbenv/versions/2.7.6/bin/bundle:23:in `load'
       15: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/gems/2.7.0/gems/bundler-2.1.4/libexec/bundle:34:in `<top (required)>'
       14: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/bundler/friendly_errors.rb:123:in `with_friendly_errors'
       13: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/gems/2.7.0/gems/bundler-2.1.4/libexec/bundle:46:in `block in <top (required)>'
       12: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/bundler/cli.rb:24:in `start'
       11: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/bundler/vendor/thor/lib/thor/base.rb:476:in `start'
       10: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/bundler/cli.rb:30:in `dispatch'
        9: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/bundler/vendor/thor/lib/thor.rb:399:in `dispatch'
        8: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/bundler/vendor/thor/lib/thor/invocation.rb:127:in `invoke_command'
        7: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/bundler/vendor/thor/lib/thor/command.rb:27:in `run'
        6: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/bundler/cli.rb:506:in `console'
        5: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/bundler/cli/console.rb:19:in `run'
        4: from (irb):1
        3: from /Users/kevin.dew/.rbenv/versions/2.7.6/lib/ruby/2.7.0/forwardable.rb:235:in `find'
        2: from /Users/kevin.dew/govuk/plek/lib/plek.rb:75:in `find'
        1: from /Users/kevin.dew/govuk/plek/lib/plek.rb:149:in `valid_service_name'
ArgumentError (Plek expects a service name only contain lowercase a-z, . (period) and - (dash) characters.)
```

This changes the behaviour so that it accepts only hostnames in
lowercase with the expected characters and doesn't try to make fixes. I
also expanded it to include numbers as this seemed an unusual omission.

The expectation is that this will have a negligible effect on real code
as it's unlikely that our committed code is not already valid.